% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LFPCA.R
\name{convert_segments_to_numeric}
\alias{convert_segments_to_numeric}
\title{Convert Segment List to Numeric}
\usage{
convert_segments_to_numeric(segment_list)
}
\arguments{
\item{segment_list}{A list where each element contains segment indices.}
}
\value{
A list of numeric vectors containing the segment indices.
}
\description{
Converts a list of segment indices (potentially stored as character or factor)
to numeric values.
}
\examples{
segments <- list(c("1", "2", "3"), c("10", "11", "12"))
convert_segments_to_numeric(segments)
}
